全面改造成若依plus的形状，有没有问题再说. 尝试重建common... 完成logout，token加入redis黑名单 完成role list查询，分页，条件查询 优化，返回菜单按照sort排序 查询字典类型 查询字典数据

+++ dit-system/src/main/java/site/ahzx/domain/entity/SysDictData.java
@@ -21,7 +21,7 @@ public class SysDictData extends BaseEntity<SysDictData> implements Serializable
      * 字典编码
      */
     @Id(keyType = KeyType.Auto)
-    private int dictCode;
+    private long dictCode;
 
     /**
      * 字典排序
